# [transformer] Faster model parallel